+++ platform/ios/platform/ios/app/MBXViewController.m
@@ -31,6 +31,11 @@ static const MGLCoordinateBounds colorado = {
     .ne = { .latitude = 40.989329, .longitude = -102.062592},
 };
 
+static const MGLCoordinateBounds areaAroundBelgium = {
+    .sw = { .latitude = 52.2782, .longitude = 8.289179999999988},
+    .ne = { .latitude = 48.5584, .longitude = 1.0162300000000073},
+};
+
 static NSString * const MBXViewControllerAnnotationViewReuseIdentifer = @"MBXViewControllerAnnotationViewReuseIdentifer";
 
 typedef NS_ENUM(NSInteger, MBXSettingsSections) {
@@ -99,6 +104,7 @@ typedef NS_ENUM(NSInteger, MBXSettingsRuntimeStylingRows) {
 };
 
 typedef NS_ENUM(NSInteger, MBXSettingsMiscellaneousRows) {
+    MBXSettingsMiscellaneousLatLngBoundsConstraints,
     MBXSettingsMiscellaneousWorldTour,
     MBXSettingsMiscellaneousRandomTour,
     MBXSettingsMiscellaneousScrollView,
@@ -110,6 +116,7 @@ typedef NS_ENUM(NSInteger, MBXSettingsMiscellaneousRows) {
     MBXSettingsMiscellaneousSetContentInsets,
     MBXSettingsMiscellaneousShowCustomLocationManager,
     MBXSettingsMiscellaneousOrnamentsPlacement,
+    MBXSettingsMiscellaneousLatLngBoundsWithPadding,
     MBXSettingsMiscellaneousPrintLogFile,
     MBXSettingsMiscellaneousDeleteLogFile
 };
@@ -227,6 +234,7 @@ CLLocationCoordinate2D randomWorldCoordinate(void) {
     BOOL _isTouringWorld;
     BOOL _contentInsetsEnabled;
     UIEdgeInsets _originalContentInsets;
+    BOOL _hasLatLngBoundConstraints;
 }
 
 #pragma mark - Setup & Teardown
@@ -426,6 +434,7 @@ CLLocationCoordinate2D randomWorldCoordinate(void) {
             break;
         case MBXSettingsMiscellaneous:
             [settingsTitles addObjectsFromArray:@[
+                _hasLatLngBoundConstraints ? @"Remove LatLng bound constraints" : @"Set LatLng bound box constraint",
                 @"Start World Tour",
                 @"Random Tour",
                 @"Embedded Map View",
@@ -437,6 +446,7 @@ CLLocationCoordinate2D randomWorldCoordinate(void) {
                 [NSString stringWithFormat:@"Turn %@ Content Insets", (_contentInsetsEnabled ? @"Off" : @"On")],
                 @"View Route Simulation",
                 @"Ornaments Placement",
+                @"Lat Long bounds with padding"
             ]];
 
             break;
@@ -645,6 +655,9 @@ CLLocationCoordinate2D randomWorldCoordinate(void) {
         case MBXSettingsMiscellaneous:
             switch (indexPath.row)
             {
+                case MBXSettingsMiscellaneousLatLngBoundsConstraints:
+                    [self setLatLngBoundsConstraints];
+                    break;
                 case MBXSettingsMiscellaneousLocalizeLabels:
                     [self toggleStyleLabelsLanguage];
                     break;
@@ -748,6 +761,9 @@ CLLocationCoordinate2D randomWorldCoordinate(void) {
                     [self.navigationController pushViewController:ornamentsViewController animated:YES];
                     break;
                 }
+                case MBXSettingsMiscellaneousLatLngBoundsWithPadding:
+                    [self flyToWithLatLngBoundsAndPadding];
+                    break;
                 default:
                     NSAssert(NO, @"All miscellaneous setting rows should be implemented");
                     break;
@@ -1451,6 +1467,20 @@ CLLocationCoordinate2D randomWorldCoordinate(void) {
     }
 }
 
+-(void)setLatLngBoundsConstraints
+{
+    if(_hasLatLngBoundConstraints) {
+        [self.mapView clearLatLnBounds];
+        [self.mapView resetPosition];
+    } else {
+        MGLMapCamera *newCamera = [self.mapView cameraThatFitsCoordinateBounds: areaAroundBelgium];
+        [self.mapView setCamera: newCamera];
+        [self.mapView setLatLngBounds: areaAroundBelgium];
+    }
+    
+    _hasLatLngBoundConstraints = !_hasLatLngBoundConstraints;
+}
+
 -(void)toggleStyleLabelsLanguage
 {
     _localizingLabels = !_localizingLabels;
@@ -1712,6 +1742,30 @@ CLLocationCoordinate2D randomWorldCoordinate(void) {
 - (UIImage *)mapView:(MGLMapView *)mapView didFailToLoadImage:(NSString *)imageName {
     UIImage *backupImage = [UIImage imageNamed:@"AppIcon"];
     return backupImage;
+}
+
+- (void)flyToWithLatLngBoundsAndPadding
+{
+    [self.mapView removeAnnotations:self.mapView.annotations];
+    CLLocationCoordinate2D sw = CLLocationCoordinate2DMake(48, 11);
+    CLLocationCoordinate2D ne = CLLocationCoordinate2DMake(50, 12);
+
+    UIEdgeInsets padding = UIEdgeInsetsMake(200, 200, 0, 0);
+    MGLMapCamera *cameraWithoutPadding = [self.mapView cameraThatFitsCoordinateBounds:MGLCoordinateBoundsMake(sw, ne)
+                                                                          edgePadding:padding];
+
+
+    MGLPointAnnotation *annotation = [MGLPointAnnotation new];
+    annotation.coordinate = cameraWithoutPadding.centerCoordinate;
+    annotation.title = @"Bounds center";
+    [self.mapView addAnnotation: annotation];
+
+    __weak MBXViewController *weakSelf = self;
+    [self.mapView flyToCamera:cameraWithoutPadding edgePadding:padding withDuration:5 completionHandler:^{
+        [weakSelf.mapView flyToCamera:cameraWithoutPadding edgePadding:UIEdgeInsetsZero withDuration:5 completionHandler:^{
+
+        }];
+    }];
 }
 
 #pragma mark - Random World Tour